A lot of things take ints in their constructors; this has been changed to doubles. Also did a lot of menial work involving game scale

+++ StarshipArena/src/Battleship.java
@@ -18,15 +18,15 @@ public class Battleship extends Starship{
 	static int primary_yoffset = -50;
 	static int primary_id = 4;
 	
-	public Battleship(StarshipArena mygame, int spawnx, int spawny){
+	public Battleship(StarshipArena mygame, double spawnx, double spawny){
 		super(mygame, spawnx, spawny);
 	}
 	
-	public Battleship(StarshipArena mygame, int spawnx, int spawny, int spawnangle){
+	public Battleship(StarshipArena mygame, double spawnx, double spawny, double spawnangle){
 		super(mygame, "none", spawnx, spawny, spawnangle);
 	}
 
-	public Battleship(StarshipArena mygame, String newteam, int spawnx, int spawny, int spawnangle){
+	public Battleship(StarshipArena mygame, String newteam, double spawnx, double spawny, double spawnangle){
 		super(mygame, newteam, spawnx, spawny, spawnangle);
 	}
 	//TODO Change stats
