+++ StarshipArena/src/Missileship.java
@@ -17,15 +17,15 @@ public class Missileship extends Starship{
 	static int primary_yoffset = -100;
 	static int primary_id = 4;
 	
-	public Missileship(StarshipArena mygame, int spawnx, int spawny){
+	public Missileship(StarshipArena mygame, double spawnx, double spawny){
 		super(mygame, spawnx, spawny);
 	}
 	
-	public Missileship(StarshipArena mygame, int spawnx, int spawny, int spawnangle){
+	public Missileship(StarshipArena mygame, double spawnx, double spawny, double spawnangle){
 		super(mygame, "none", spawnx, spawny, spawnangle);
 	}
 
-	public Missileship(StarshipArena mygame, String newteam, int spawnx, int spawny, int spawnangle){
+	public Missileship(StarshipArena mygame, String newteam, double spawnx, double spawny, double spawnangle){
 		super(mygame, newteam, spawnx, spawny, spawnangle);
 	}
